+++ fft.h
@@ -1,8 +1,8 @@
-void prepare_fft(long long int sz);
+void prepare_fft(int sz);
 void free_fft();
-void fft(double complex *in, double complex *out, long long int sz);
+void fft(double complex *in, double complex *out, int sz);
 
-void fft2(double complex *in, double complex *out, long long int sz);
+void fft2(double complex *in, double complex *out, int sz);
 
-void fft3(double complex *in, double complex *out, long long int sz);
+void fft3(double complex *in, double complex *out, int sz);
